:recycle: refactor: QueryLoader 패키기 경로 수정
